Reapply "dbeaver/pro#8476 copilot responses api & OpenAI client refactoring (#41056)"
This reverts commit ce460acc1730e2c853cbac6a014ee7550c8a87db.
