+++ plugins/org.jkiss.dbeaver.model.ai/src/org/jkiss/dbeaver/model/ai/engine/copilot/CopilotCompletionEngine.java
@@ -17,35 +17,38 @@
 package org.jkiss.dbeaver.model.ai.engine.copilot;
 
 import org.jkiss.code.NotNull;
-import org.jkiss.code.Nullable;
 import org.jkiss.dbeaver.DBException;
+import org.jkiss.dbeaver.model.ai.AIFunctionCall;
 import org.jkiss.dbeaver.model.ai.AIMessage;
 import org.jkiss.dbeaver.model.ai.AIMessageType;
 import org.jkiss.dbeaver.model.ai.AIUsage;
 import org.jkiss.dbeaver.model.ai.engine.*;
 import org.jkiss.dbeaver.model.ai.engine.copilot.dto.*;
 import org.jkiss.dbeaver.model.ai.engine.openai.OpenAIConstants;
+import org.jkiss.dbeaver.model.ai.engine.openai.OpenAiUtils;
+import org.jkiss.dbeaver.model.ai.engine.openai.dto.OAIMessage;
+import org.jkiss.dbeaver.model.ai.engine.openai.dto.OAIResponsesResponse;
 import org.jkiss.dbeaver.model.ai.engine.openai.dto.OAITool;
 import org.jkiss.dbeaver.model.ai.internal.AIMessages;
 import org.jkiss.dbeaver.model.ai.utils.DisposableLazyValue;
 import org.jkiss.dbeaver.model.runtime.DBRProgressMonitor;
 import org.jkiss.utils.CommonUtils;
 
+import java.util.ArrayList;
 import java.util.List;
-import java.util.Objects;
 import java.util.Set;
 
 public class CopilotCompletionEngine<P extends CopilotProperties> extends BaseCompletionEngine<P> {
 
-    protected final DisposableLazyValue<CopilotClient, DBException> client = new DisposableLazyValue<>() {
+    protected final DisposableLazyValue<CopilotClientResponses, DBException> client = new DisposableLazyValue<>() {
         @NotNull
         @Override
-        protected CopilotClient initialize() throws DBException {
+        protected CopilotClientResponses initialize() throws DBException {
             return createClient(getProperties().getBaseAuthUrl());
         }
 
         @Override
-        protected void onDispose(@NotNull CopilotClient disposedValue) {
+        protected void onDispose(@NotNull CopilotClientResponses disposedValue) {
             disposedValue.close();
         }
     };
@@ -58,11 +61,13 @@ public class CopilotCompletionEngine<P extends CopilotProperties> extends BaseCo
     @NotNull
     @Override
     public List<AIModel> getModels(@NotNull DBRProgressMonitor monitor) throws DBException {
-        return client.getInstance().loadModels(monitor, requestSessionToken(monitor).token()).stream()
-            .map(model -> CopilotModels.getModelByName(model.id()).orElse(
-                new AIModel(model.id(), null, Set.of())
-            ))
-            .toList();
+        List<CopilotModel> copilotModels = client.getInstance().loadModels(monitor, requestSessionToken(monitor).token());
+        List<AIModel> list = new ArrayList<>();
+        for (CopilotModel model : copilotModels) {
+            AIModel aiModel = new AIModel(model.id(), null, Set.of(AIModelFeature.CHAT));
+            list.add(aiModel);
+        }
+        return list;
     }
 
     @NotNull
@@ -71,10 +76,11 @@ public class CopilotCompletionEngine<P extends CopilotProperties> extends BaseCo
         @NotNull DBRProgressMonitor monitor,
         @NotNull AIEngineRequest request
     ) throws DBException {
-        CopilotChatResponse chatResponse = client.getInstance().chat(
+        OAIResponsesResponse chatResponse = client.getInstance().chat(
             monitor,
             requestSessionToken(monitor).token(),
-            createChatRequest(request, false)
+            createLegacyChatRequest(request, false),
+            OpenAiUtils.createOpenAiRequest(request, getModelName(), getProperties().getTemperature())
         );
 
         return toEngineResponse(chatResponse);
@@ -89,7 +95,8 @@ public class CopilotCompletionEngine<P extends CopilotProperties> extends BaseCo
         client.getInstance().createChatCompletionStream(
             monitor,
             requestSessionToken(monitor).token(),
-            createChatRequest(request, true),
+            OpenAiUtils.createOpenAiRequest(request, getModelName(), null),
+            createLegacyChatRequest(request, true),
             listener
         );
     }
@@ -124,7 +131,8 @@ public class CopilotCompletionEngine<P extends CopilotProperties> extends BaseCo
         return sessionToken;
     }
 
-    public String getModelName() throws DBException {
+    @NotNull
+    public String getModelName() {
         return CommonUtils.toString(
             properties.getModel(),
             OpenAIConstants.DEFAULT_MODEL
@@ -132,10 +140,10 @@ public class CopilotCompletionEngine<P extends CopilotProperties> extends BaseCo
     }
 
     @NotNull
-    private CopilotChatRequest createChatRequest(
+    private CopilotChatRequest createLegacyChatRequest(
         @NotNull AIEngineRequest request,
         boolean stream
-    ) throws DBException {
+    ) {
         return CopilotChatRequest.builder()
             .withModel(getModelName())
             .withMessages(toCopilotMessages(request.getMessages()))
@@ -151,36 +159,31 @@ public class CopilotCompletionEngine<P extends CopilotProperties> extends BaseCo
             .build();
     }
 
+
     @NotNull
-    private AIEngineResponse toEngineResponse(@NotNull CopilotChatResponse response) throws DBException {
-        AIUsage usage = response.getAIUsage();
-        CopilotChatResponse.ToolCall toolCall = getFirstToolCall(response);
-        if (toolCall != null) {
-            return new AIEngineResponse(CopilotUtils.createFunctionCall(toolCall), usage);
-        }
-
-        List<String> variants = response.choices().stream()
-            .map(CopilotChatResponse.Choice::message)
-            .map(CopilotChatResponse.Message::content)
-            .filter(CommonUtils::isNotEmpty)
+    private AIEngineResponse toEngineResponse(@NotNull OAIResponsesResponse response) throws DBException {
+        List<OAIMessage> messages = response.output.stream()
+            .filter(msg -> !OAIMessage.TYPE_FUNCTION_REASONING.equals(msg.type))
             .toList();
-        if (variants.isEmpty()) {
-            variants = List.of(AIMessages.ai_empty_engine_response);
+        AIUsage usage = response.getAIUsage();
+        if (messages.isEmpty()) {
+            return new AIEngineResponse(
+                AIMessageType.ASSISTANT,
+                List.of(AIMessages.ai_empty_engine_response),
+                usage
+            );
         }
+        OAIMessage message = messages.getFirst();
+        if (OAIMessage.TYPE_FUNCTION_CALL.equals(message.type)) {
+            AIFunctionCall fc = OpenAiUtils.createFunctionCall(message);
+            return new AIEngineResponse(fc, usage);
+        } else {
+            List<String> choices = messages.stream()
+                .map(OAIMessage::getFullText)
+                .toList();
 
-        return new AIEngineResponse(AIMessageType.ASSISTANT, variants, usage);
-    }
-
-    @Nullable
-    private static CopilotChatResponse.ToolCall getFirstToolCall(@NotNull CopilotChatResponse response) {
-        return response.choices().stream()
-            .map(CopilotChatResponse.Choice::message)
-            .filter(Objects::nonNull)
-            .map(CopilotChatResponse.Message::toolCalls)
-            .filter(calls -> calls != null && !calls.isEmpty())
-            .map(List::getFirst)
-            .findFirst()
-            .orElse(null);
+            return new AIEngineResponse(AIMessageType.ASSISTANT, choices, usage);
+        }
     }
 
     @NotNull
@@ -191,12 +194,12 @@ public class CopilotCompletionEngine<P extends CopilotProperties> extends BaseCo
     }
 
     @NotNull
-    protected CopilotClient createClient(@NotNull String baseAuthUrl) throws DBException {
+    protected CopilotClientResponses createClient(@NotNull String baseAuthUrl) throws DBException {
         String token = properties.getToken();
         if (token == null || token.isEmpty()) {
             throw new DBException("Copilot API token is not set");
         }
 
-        return new CopilotClient(baseAuthUrl);
+        return new CopilotClientResponses(baseAuthUrl);
     }
 }
