+++ plugins/org.jkiss.dbeaver.model.ai/src/org/jkiss/dbeaver/model/ai/engine/copilot/dto/CopilotChatResponseLegacy.java
@@ -24,7 +24,7 @@ import org.jkiss.dbeaver.model.ai.AIUsage;
 import java.util.List;
 import java.util.Map;
 
-public record CopilotChatResponse(
+public record CopilotChatResponseLegacy(
     @Nullable
     String id,
     @NotNull
