+++ plugins/org.jkiss.dbeaver.model.ai/src/org/jkiss/dbeaver/model/ai/utils/MonitoredHttpClient.java
@@ -1,6 +1,6 @@
 /*
  * DBeaver - Universal Database Manager
- * Copyright (C) 2010-2025 DBeaver Corp and others
+ * Copyright (C) 2010-2026 DBeaver Corp and others
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,6 +17,7 @@
 package org.jkiss.dbeaver.model.ai.utils;
 
 import org.jkiss.code.NotNull;
+import org.jkiss.code.Nullable;
 import org.jkiss.dbeaver.DBException;
 import org.jkiss.dbeaver.model.runtime.DBRProgressMonitor;
 
@@ -27,9 +28,10 @@ import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.TimeUnit;
 import java.util.function.Consumer;
-import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 public class MonitoredHttpClient implements AutoCloseable {
+
     /**
      * Maps an HTTP status code and response body to a {@link DBException}.
      */
@@ -39,12 +41,28 @@ public class MonitoredHttpClient implements AutoCloseable {
         DBException map(int statusCode, @NotNull String body);
     }
 
+    /**
+     * Handles errors that occur during the processing of an HTTP response.
+     */
+    @FunctionalInterface
+    public interface ErrorProcessor {
+        boolean process(
+            @NotNull ErrorMapper mapper,
+            @NotNull Consumer<Throwable> errorHandler,
+            @NotNull HttpResponse<Stream<String>> response,
+            @Nullable Runnable backupOption,
+            int statusCode
+        );
+    }
+
     private final HttpClient client;
     private final ErrorMapper errorMapper;
+    private final ErrorProcessor errorProcessor;
 
-    public MonitoredHttpClient(@NotNull HttpClient client, @NotNull ErrorMapper errorMapper) {
+    public MonitoredHttpClient(@NotNull HttpClient client, @NotNull ErrorMapper errorMapper, @NotNull ErrorProcessor processor) {
         this.client = client;
         this.errorMapper = errorMapper;
+        this.errorProcessor = processor;
     }
 
     @NotNull
@@ -59,8 +77,8 @@ public class MonitoredHttpClient implements AutoCloseable {
      */
     @NotNull
     public String send(
-        DBRProgressMonitor monitor,
-        HttpRequest request
+        @NotNull DBRProgressMonitor monitor,
+        @NotNull HttpRequest request
     ) throws DBException {
         monitor.beginTask("Request AI completion", 1);
 
@@ -96,18 +114,28 @@ public class MonitoredHttpClient implements AutoCloseable {
         }
     }
 
+    @NotNull
     public CompletableFuture<Void> sendAsync(
         @NotNull HttpRequest request,
         @NotNull Consumer<String> eventHandler,
         @NotNull Consumer<Throwable> errorHandler,
         @NotNull Runnable completionHandler
     ) {
+        return sendAsync(request, eventHandler, errorHandler, completionHandler, null);
+    }
+
+    @NotNull
+    public CompletableFuture<Void> sendAsync(
+        @NotNull HttpRequest request,
+        @NotNull Consumer<String> eventHandler,
+        @NotNull Consumer<Throwable> errorHandler,
+        @NotNull Runnable completionHandler,
+        @Nullable Runnable backupOption
+    ) {
         return client.sendAsync(request, HttpResponse.BodyHandlers.ofLines())
             .thenAccept(response -> {
                 int statusCode = response.statusCode();
-                if (statusCode != 200) {
-                    String responseBody = response.body().collect(Collectors.joining());
-                    errorHandler.accept(errorMapper.map(statusCode, responseBody));
+                if (errorProcessor.process(errorMapper, errorHandler, response, backupOption, statusCode)) {
                     return;
                 }
