+++ plugins/org.jkiss.dbeaver.ui.ai/src/org/jkiss/dbeaver/ui/ai/engine/openai/OpenAiConfigurator.java
@@ -1,6 +1,6 @@
 /*
  * DBeaver - Universal Database Manager
- * Copyright (C) 2010-2025 DBeaver Corp and others
+ * Copyright (C) 2010-2026 DBeaver Corp and others
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -33,7 +33,7 @@ import org.jkiss.dbeaver.DBException;
 import org.jkiss.dbeaver.model.ai.engine.AIEngineProperties;
 import org.jkiss.dbeaver.model.ai.engine.AIModel;
 import org.jkiss.dbeaver.model.ai.engine.AIModelFeature;
-import org.jkiss.dbeaver.model.ai.engine.openai.OpenAIClient;
+import org.jkiss.dbeaver.model.ai.engine.openai.OpenAIClientResponses;
 import org.jkiss.dbeaver.model.ai.engine.openai.OpenAIEngine;
 import org.jkiss.dbeaver.model.ai.engine.openai.OpenAIModels;
 import org.jkiss.dbeaver.model.ai.engine.openai.OpenAIProperties;
@@ -69,7 +69,6 @@ public class OpenAiConfigurator<ENGINE extends AIEngineDescriptor, PROPERTIES ex
     private Text temperatureText;
     private ModelSelectorField modelSelectorField;
     private ContextWindowSizeField contextWindowSizeField;
-    private Button legacyApiCheck;
     private Button logQueryCheck;
 
     protected final CachedValue<List<AIModel>> modelsCache = new CachedValue<>(this::fetchOpenAiModels);
@@ -77,7 +76,7 @@ public class OpenAiConfigurator<ENGINE extends AIEngineDescriptor, PROPERTIES ex
     @Override
     public void createControl(
         @NotNull Composite parent,
-        AIEngineDescriptor object,
+        @NotNull AIEngineDescriptor object,
         @NotNull Runnable propertyChangeListener
     ) {
         Composite composite = UIUtils.createComposite(parent, 3);
@@ -94,7 +93,7 @@ public class OpenAiConfigurator<ENGINE extends AIEngineDescriptor, PROPERTIES ex
     public void loadSettings(@NotNull PROPERTIES configuration) {
         baseUrl = CommonUtils.toString(configuration.getBaseUrl());
         if (baseUrl.isEmpty()) {
-            baseUrl = OpenAIClient.OPENAI_ENDPOINT;
+            baseUrl = OpenAIClientResponses.OPENAI_ENDPOINT;
         }
         token = CommonUtils.toString(configuration.getToken());
         modelSelectorField.setSelectedModel(
@@ -102,7 +101,6 @@ public class OpenAiConfigurator<ENGINE extends AIEngineDescriptor, PROPERTIES ex
         );
         temperature = CommonUtils.toString(configuration.getTemperature(), "0.0");
 
-        useLegacyApi = configuration.isLegacyApi();
         logQuery = CommonUtils.toBoolean(configuration.isLoggingEnabled());
         applySettings();
 
@@ -118,7 +116,6 @@ public class OpenAiConfigurator<ENGINE extends AIEngineDescriptor, PROPERTIES ex
         configuration.setModel(modelSelectorField.getSelectedModel());
         configuration.setContextWindowSize(contextWindowSizeField.getValue());
         configuration.setTemperature(CommonUtils.toDouble(temperature));
-        configuration.setLegacyApi(useLegacyApi);
         configuration.setLoggingEnabled(logQuery);
     }
 
@@ -128,16 +125,6 @@ public class OpenAiConfigurator<ENGINE extends AIEngineDescriptor, PROPERTIES ex
     }
 
     protected void createAdditionalSettings(@NotNull Composite parent) {
-        legacyApiCheck = UIUtils.createCheckbox(
-            parent,
-            AIUIMessages.openai_configurator_legacy_api_label,
-            AIUIMessages.openai_configurator_legacy_api_tip,
-            false,
-            2
-        );
-        legacyApiCheck.addSelectionListener(SelectionListener.widgetSelectedAdapter(e ->
-            useLegacyApi = legacyApiCheck.getSelection()
-        ));
         logQueryCheck = UIUtils.createCheckbox(
             parent,
             AIUIMessages.openai_configurator_log_query_label,
@@ -187,7 +174,8 @@ public class OpenAiConfigurator<ENGINE extends AIEngineDescriptor, PROPERTIES ex
         temperatureText.addModifyListener((e) -> temperature = temperatureText.getText());
     }
 
-    private List<AIModel> fetchOpenAiModels(DBRProgressMonitor monitor) throws DBException {
+    @NotNull
+    private List<AIModel> fetchOpenAiModels(@NotNull DBRProgressMonitor monitor) throws DBException {
         if (token == null || token.isEmpty()) {
             throw new DBException(AIUIMessages.openai_configurator_token_required);
         }
@@ -234,7 +222,7 @@ public class OpenAiConfigurator<ENGINE extends AIEngineDescriptor, PROPERTIES ex
             NLS.bind(AIUIMessages.gpt_preference_page_token_info, getApiKeyURL()),
             new SelectionAdapter() {
                 @Override
-                public void widgetSelected(SelectionEvent e) {
+                public void widgetSelected(@NotNull SelectionEvent e) {
                     UIUtils.openWebBrowser(getApiKeyURL());
                 }
             }
@@ -244,6 +232,7 @@ public class OpenAiConfigurator<ENGINE extends AIEngineDescriptor, PROPERTIES ex
         link.setLayoutData(gd);
     }
 
+    @NotNull
     protected String getApiKeyURL() {
         return API_KEY_URL;
     }
@@ -257,7 +246,6 @@ public class OpenAiConfigurator<ENGINE extends AIEngineDescriptor, PROPERTIES ex
         }
 
         temperatureText.setText(temperature);
-        legacyApiCheck.setSelection(useLegacyApi);
         logQueryCheck.setSelection(logQuery);
     }
 
@@ -269,6 +257,7 @@ public class OpenAiConfigurator<ENGINE extends AIEngineDescriptor, PROPERTIES ex
     }
 
     @Override
+    @NotNull
     public Optional<AIEngineProperties> getCurrentProperties() {
         OpenAIProperties propertiesCopy = new OpenAIProperties();
         propertiesCopy.setBaseUrl(baseUrl);
@@ -276,7 +265,6 @@ public class OpenAiConfigurator<ENGINE extends AIEngineDescriptor, PROPERTIES ex
         propertiesCopy.setModel(modelSelectorField.getSelectedModel());
         propertiesCopy.setContextWindowSize(contextWindowSizeField.getValue());
         propertiesCopy.setTemperature(CommonUtils.toDouble(temperature));
-        propertiesCopy.setLegacyApi(useLegacyApi);
         propertiesCopy.setLoggingEnabled(logQuery);
         return Optional.of(propertiesCopy);
     }
